Track fetch timeouts per backend request
Instead of tracking the fetch timeout globally this passes it to the
urllib3 backend calls, so the timeouts are per call to each backend.
Furthermore there's a MAX_FETCH_RETRIES setting. So the maximum number
of seconds that we will wait for a given backend is:
MAX_FETCH_RETRIES * REMOTE_FETCH_TIMEOUT

If USE_WORKER_POOL is False then the backends are queried in series, so
we'll wait for one backend after the other. In this case the max number
of seconds we'll wait for all backends is:
MAX_FETCH_RETRIES * REMOTE_FETCH_TIMEOUT * len(CLUSTER_SERVERS)